table: do not merge content cells with empty ones (#280)
