+++ list/render_markdown.go
@@ -1,15 +1,15 @@
 package list
 
 // RenderMarkdown renders the List in the Markdown format. Example:
-//    * Game Of Thrones
-//      * Winter
-//      * Is
-//      * Coming
-//        * This
-//        * Is
-//        * Known
-//    * The Dark Tower
-//      * The Gunslinger
+// | * Game Of Thrones
+// |    * Winter
+// |    * Is
+// |    * Coming
+// |      * This
+// |      * Is
+// |      * Known
+// |  * The Dark Tower
+// |    * The Gunslinger
 func (l *List) RenderMarkdown() string {
 	// make a copy of the original style and ensure it is restored on exit
 	originalStyle := l.style
