+++ sonar-project.properties
@@ -1,0 +1,9 @@
+# Define the root directory for sources and tests
+sonar.sources = ./
+sonar.tests = ./
+
+# Include test files in test scope
+sonar.test.inclusions = **/*_test.go
+
+# Exclude test files from source scope
+sonar.exclusions = **/*_test.go
