+++ table/render_html.go
@@ -13,49 +13,50 @@ const (
 )
 
 // RenderHTML renders the Table in HTML format. Example:
-//  <table class="go-pretty-table">
-//    <thead>
-//    <tr>
-//      <th align="right">#</th>
-//      <th>First Name</th>
-//      <th>Last Name</th>
-//      <th align="right">Salary</th>
-//      <th>&nbsp;</th>
-//    </tr>
-//    </thead>
-//    <tbody>
-//    <tr>
-//      <td align="right">1</td>
-//      <td>Arya</td>
-//      <td>Stark</td>
-//      <td align="right">3000</td>
-//      <td>&nbsp;</td>
-//    </tr>
-//    <tr>
-//      <td align="right">20</td>
-//      <td>Jon</td>
-//      <td>Snow</td>
-//      <td align="right">2000</td>
-//      <td>You know nothing, Jon Snow!</td>
-//    </tr>
-//    <tr>
-//      <td align="right">300</td>
-//      <td>Tyrion</td>
-//      <td>Lannister</td>
-//      <td align="right">5000</td>
-//      <td>&nbsp;</td>
-//    </tr>
-//    </tbody>
-//    <tfoot>
-//    <tr>
-//      <td align="right">&nbsp;</td>
-//      <td>&nbsp;</td>
-//      <td>Total</td>
-//      <td align="right">10000</td>
-//      <td>&nbsp;</td>
-//    </tr>
-//    </tfoot>
-//  </table>
+//
+//	<table class="go-pretty-table">
+//	  <thead>
+//	  <tr>
+//	    <th align="right">#</th>
+//	    <th>First Name</th>
+//	    <th>Last Name</th>
+//	    <th align="right">Salary</th>
+//	    <th>&nbsp;</th>
+//	  </tr>
+//	  </thead>
+//	  <tbody>
+//	  <tr>
+//	    <td align="right">1</td>
+//	    <td>Arya</td>
+//	    <td>Stark</td>
+//	    <td align="right">3000</td>
+//	    <td>&nbsp;</td>
+//	  </tr>
+//	  <tr>
+//	    <td align="right">20</td>
+//	    <td>Jon</td>
+//	    <td>Snow</td>
+//	    <td align="right">2000</td>
+//	    <td>You know nothing, Jon Snow!</td>
+//	  </tr>
+//	  <tr>
+//	    <td align="right">300</td>
+//	    <td>Tyrion</td>
+//	    <td>Lannister</td>
+//	    <td align="right">5000</td>
+//	    <td>&nbsp;</td>
+//	  </tr>
+//	  </tbody>
+//	  <tfoot>
+//	  <tr>
+//	    <td align="right">&nbsp;</td>
+//	    <td>&nbsp;</td>
+//	    <td>Total</td>
+//	    <td align="right">10000</td>
+//	    <td>&nbsp;</td>
+//	  </tr>
+//	  </tfoot>
+//	</table>
 func (t *Table) RenderHTML() string {
 	t.initForRender()
