+++ table/render_markdown.go
@@ -6,12 +6,13 @@ import (
 )
 
 // RenderMarkdown renders the Table in Markdown format. Example:
-//  | # | First Name | Last Name | Salary |  |
-//  | ---:| --- | --- | ---:| --- |
-//  | 1 | Arya | Stark | 3000 |  |
-//  | 20 | Jon | Snow | 2000 | You know nothing, Jon Snow! |
-//  | 300 | Tyrion | Lannister | 5000 |  |
-//  |  |  | Total | 10000 |  |
+//
+//	| # | First Name | Last Name | Salary |  |
+//	| ---:| --- | --- | ---:| --- |
+//	| 1 | Arya | Stark | 3000 |  |
+//	| 20 | Jon | Snow | 2000 | You know nothing, Jon Snow! |
+//	| 300 | Tyrion | Lannister | 5000 |  |
+//	|  |  | Total | 10000 |  |
 func (t *Table) RenderMarkdown() string {
 	t.initForRender()
