+++ text/ansi.go
@@ -8,7 +8,8 @@ var ANSICodesSupported = areANSICodesSupported()
 
 // Escape encodes the string with the ANSI Escape Sequence.
 // For ex.:
-//  Escape("Ghost", "") == "Ghost"
+//
+//	Escape("Ghost", "") == "Ghost"
 //	Escape("Ghost", "\x1b[91m") == "\x1b[91mGhost\x1b[0m"
 //	Escape("\x1b[94mGhost\x1b[0mLady", "\x1b[91m") == "\x1b[94mGhost\x1b[0m\x1b[91mLady\x1b[0m"
 //	Escape("Nymeria\x1b[94mGhost\x1b[0mLady", "\x1b[91m") == "\x1b[91mNymeria\x1b[94mGhost\x1b[0m\x1b[91mLady\x1b[0m"
@@ -30,7 +31,8 @@ func Escape(str string, escapeSeq string) string {
 
 // StripEscape strips all ANSI Escape Sequence from the string.
 // For ex.:
-//  StripEscape("Ghost") == "Ghost"
+//
+//	StripEscape("Ghost") == "Ghost"
 //	StripEscape("\x1b[91mGhost\x1b[0m") == "Ghost"
 //	StripEscape("\x1b[94mGhost\x1b[0m\x1b[91mLady\x1b[0m") == "GhostLady"
 //	StripEscape("\x1b[91mNymeria\x1b[94mGhost\x1b[0m\x1b[91mLady\x1b[0m") == "NymeriaGhostLady"
