+++ text/valign.go
@@ -14,12 +14,12 @@ const (
 )
 
 // Apply aligns the lines vertically. For ex.:
-//  * VAlignTop.Apply({"Game", "Of", "Thrones"},    5)
-// 	    returns {"Game", "Of", "Thrones", "", ""}
-//  * VAlignMiddle.Apply({"Game", "Of", "Thrones"}, 5)
-// 	    returns {"", "Game", "Of", "Thrones", ""}
-//  * VAlignBottom.Apply({"Game", "Of", "Thrones"}, 5)
-// 	    returns {"", "", "Game", "Of", "Thrones"}
+//   - VAlignTop.Apply({"Game", "Of", "Thrones"},    5)
+//     returns {"Game", "Of", "Thrones", "", ""}
+//   - VAlignMiddle.Apply({"Game", "Of", "Thrones"}, 5)
+//     returns {"", "Game", "Of", "Thrones", ""}
+//   - VAlignBottom.Apply({"Game", "Of", "Thrones"}, 5)
+//     returns {"", "", "Game", "Of", "Thrones"}
 func (va VAlign) Apply(lines []string, maxLines int) []string {
 	if len(lines) == maxLines {
 		return lines
@@ -42,12 +42,12 @@ func (va VAlign) Apply(lines []string, maxLines int) []string {
 }
 
 // ApplyStr aligns the string (of 1 or more lines) vertically. For ex.:
-//  * VAlignTop.ApplyStr("Game\nOf\nThrones",    5)
-// 	    returns {"Game", "Of", "Thrones", "", ""}
-//  * VAlignMiddle.ApplyStr("Game\nOf\nThrones", 5)
-// 	    returns {"", "Game", "Of", "Thrones", ""}
-//  * VAlignBottom.ApplyStr("Game\nOf\nThrones", 5)
-// 	    returns {"", "", "Game", "Of", "Thrones"}
+//   - VAlignTop.ApplyStr("Game\nOf\nThrones",    5)
+//     returns {"Game", "Of", "Thrones", "", ""}
+//   - VAlignMiddle.ApplyStr("Game\nOf\nThrones", 5)
+//     returns {"", "Game", "Of", "Thrones", ""}
+//   - VAlignBottom.ApplyStr("Game\nOf\nThrones", 5)
+//     returns {"", "", "Game", "Of", "Thrones"}
 func (va VAlign) ApplyStr(text string, maxLines int) []string {
 	return va.Apply(strings.Split(text, "\n"), maxLines)
 }
